Add files via upload
The weather station uses esp8266 to connect to the Ubidot IOT platform. 
Sensors used: 
1. DHT11 for  measuring humidity and temperature 
2. Rain sensor to detect the presence of rain
Operations: 
Dew point was calculated from the humidity and temperature values. 
The due point can be compared to the area due point to predict the possibility of rain 
The data generated can be used to determine the accuracy of the station by comparing the rain prediction vs if it actually rains